Move webgl rect integration test to new rect.ts
