+++ bokehjs/test/baselines/linux/Rect_glyph__should_support_hatch_patterns_and_line_joins.blf
