+++ bokehjs/test/baselines/windows/Rect_glyph__should_support_hatch_patterns_and_line_joins.png
